Move playgound to components and change icons
